refactor(component): 함수형컴포넌트를 함수정의-> 화살표함수 방식으로 고침
