+++ src/components/ProductList/ProductCard/Stepper/Stepper.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Stepper = () => {
+  return <div>Stepper</div>;
+};
+
+export default Stepper;
